Move protobuf-returning functions into their own file
